Move the Repository update job into it's own class
